LayoutTests/fast/regions/cssom/get-region-flow-ranges-absolute-pos-expected.txt

[CSSRegions] Test for Region::getRegionFlowRanges for absolute positioned elements
PASS range1.length is 1
PASS Array [wrapper,0,wrapper,1] is equal to [wrapper,0,wrapper,1]
PASS successfullyParsed is true

TEST COMPLETE
